Reduce duplicte code
