+++ GroundStation/MQTT/SendToMqtt.cs
@@ -10,9 +10,7 @@
 
 #endregion "copyright"
 
-using MQTTnet;
-using MQTTnet.Client.Disconnecting;
-using MQTTnet.Client.Options;
+using DaleGhent.NINA.GroundStation.Mqtt;
 using Newtonsoft.Json;
 using NINA.Core.Model;
 using NINA.Core.Utility;
@@ -21,9 +19,7 @@ using NINA.Sequencer.Validations;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.ComponentModel;
 using System.ComponentModel.Composition;
-using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -36,20 +32,14 @@ namespace DaleGhent.NINA.GroundStation.SendToMqtt {
     [Export(typeof(ISequenceItem))]
     [JsonObject(MemberSerialization.OptIn)]
     public class SendToMqtt : SequenceItem, IValidatable {
+        private MqttCommon mqtt;
         private string topic;
         private string payload = string.Empty;
 
         [ImportingConstructor]
         public SendToMqtt() {
-            MqttBrokerHost = Properties.Settings.Default.MqttBrokerHost;
-            MqttBrokerPort = Properties.Settings.Default.MqttBrokerPort;
-            MqttBrokerUseTls = Properties.Settings.Default.MqttBrokerUseTls;
-            MqttUsername = Security.Decrypt(Properties.Settings.Default.MqttUsername);
-            MqttPassword = Security.Decrypt(Properties.Settings.Default.MqttPassword);
-            MqttClientId = Properties.Settings.Default.MqttClientId;
+            mqtt = new MqttCommon();
             Topic = Properties.Settings.Default.MqttDefaultTopic;
-
-            Properties.Settings.Default.PropertyChanged += SettingsChanged;
         }
 
         public SendToMqtt(SendToMqtt copyMe) : this() {
@@ -75,58 +65,14 @@ namespace DaleGhent.NINA.GroundStation.SendToMqtt {
         }
 
         public override async Task Execute(IProgress<ApplicationStatus> progress, CancellationToken ct) {
-            var factory = new MqttFactory();
-            var mqttClient = factory.CreateMqttClient();
-
-            var options = new MqttClientOptionsBuilder()
-                .WithClientId(MqttClientId)
-                .WithTcpServer(MqttBrokerHost, MqttBrokerPort)
-                .WithCleanSession();
-
-            if (!string.IsNullOrEmpty(MqttUsername) && !string.IsNullOrWhiteSpace(MqttUsername)) {
-                options.WithCredentials(MqttUsername, MqttPassword);
-            }
-
-            if (MqttBrokerUseTls) {
-                options.WithTls();
-            }
-
-            var opts = options.Build();
-
-            var message = new MqttApplicationMessageBuilder()
-                .WithTopic(Topic)
-                .WithPayload(Payload)
-                .WithExactlyOnceQoS()
-                .WithRetainFlag()
-                .Build();
-
-            var discopts = new MqttClientDisconnectOptions();
-
-            Logger.Debug("SendToMqtt: Pushing message");
-
-            try {
-                await mqttClient.ConnectAsync(opts, ct);
-                await mqttClient.PublishAsync(message, ct);
-                await mqttClient.DisconnectAsync(discopts, ct);
-                mqttClient.Dispose();
-            } catch (Exception ex) {
-                Logger.Error($"Error sending to MQTT broker: {ex.Message}");
-                throw ex;
-            }
+            Logger.Trace($"{this}: {Payload}");
+            await mqtt.PublishMessge(Topic, Payload, ct);
         }
 
         public IList<string> Issues { get; set; } = new ObservableCollection<string>();
 
         public bool Validate() {
-            var i = new List<string>();
-
-            if (string.IsNullOrEmpty(MqttBrokerHost) || string.IsNullOrWhiteSpace(MqttBrokerHost)) {
-                i.Add("MQTT broker hostname or IP not configured");
-            }
-
-            if (string.IsNullOrEmpty(MqttClientId) || string.IsNullOrWhiteSpace(MqttClientId)) {
-                i.Add("MQTT client ID is invalid!");
-            }
+            var i = new List<string>(mqtt.ValidateSettings());
 
             if (i != Issues) {
                 Issues = i;
@@ -149,36 +95,6 @@ namespace DaleGhent.NINA.GroundStation.SendToMqtt {
 
         public override string ToString() {
             return $"Category: {Category}, Item: {nameof(SendToMqtt)}";
-        }
-
-        private string MqttBrokerHost { get; set; }
-        private ushort MqttBrokerPort { get; set; }
-        private bool MqttBrokerUseTls { get; set; }
-        private string MqttUsername { get; set; }
-        private string MqttPassword { get; set; }
-        private string MqttClientId { get; set; }
-
-        private void SettingsChanged(object sender, PropertyChangedEventArgs e) {
-            switch (e.PropertyName) {
-                case "MqttBrokerHost":
-                    MqttBrokerHost = Properties.Settings.Default.MqttBrokerHost;
-                    break;
-                case "MqttBrokerPort":
-                    MqttBrokerPort = Properties.Settings.Default.MqttBrokerPort;
-                    break;
-                case "MqttBrokerUseTls":
-                    MqttBrokerUseTls = Properties.Settings.Default.MqttBrokerUseTls;
-                    break;
-                case "MqttUsername":
-                    MqttUsername = Security.Decrypt(Properties.Settings.Default.MqttUsername);
-                    break;
-                case "MqttPassword":
-                    MqttPassword = Security.Decrypt(Properties.Settings.Default.MqttPassword);
-                    break;
-                case "MqttClientId":
-                    MqttClientId = Properties.Settings.Default.MqttClientId;
-                    break;
-            }
         }
     }
 }
